Move timeout controllers under subdirectory
Also rename modules and commands to be more explicit (using "set"/"list"
prefixes).
